autotest: Embed gdaltestaux into gdaltest.
It is no longer necessary to place this code in a separate file, since
there's only one copy of it for Python 3 only.
